+++ src/cookie_agent/environment/__init__.py
@@ -1,0 +1,12 @@
+"""Public Environment package exports."""
+
+from cookie_agent.environment.environment import CookieEnvironment
+from cookie_agent.environment.exceptions import CaptureTimeoutError, EnvironmentError
+from cookie_agent.environment.info import StepInfo
+
+__all__: list[str] = [
+    "CookieEnvironment",
+    "EnvironmentError",
+    "CaptureTimeoutError",
+    "StepInfo",
+]